+++ crates/app/pages/repos/_/models/_/production_alerts/new/server/Cargo.toml
@@ -19,6 +19,7 @@ num = "0.4"
 pinwheel = "0.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+sqlx = { version = "0.5", default-features = false, features = ["any", "chrono", "postgres", "runtime-tokio-rustls", "sqlite"] }
 sunfish = "0.2"
 tangram_app_common = { path = "../../../../../../../../common" }
 tangram_app_layouts = { path = "../../../../../../../../layouts" }
