+++ crates/app/pages/repos/_/models/_/production_alerts/new/server/post.rs
@@ -4,9 +4,9 @@ use multer::Multipart;
 use pinwheel::prelude::*;
 use std::sync::Arc;
 use tangram_app_common::{
+	alerts::{AlertHeuristics, AlertThreshold},
 	error::{bad_request, not_found, redirect_to_login, service_unavailable},
 	path_components,
-	repos::add_model_version,
 	user::{authorize_user, authorize_user_for_model, authorize_user_for_repo},
 	Context,
 };
@@ -77,7 +77,9 @@ pub async fn post(request: &mut http::Request<hyper::Body>) -> Result<http::Resp
 			return Ok(response);
 		}
 	};
-	let mut file: Option<Vec<u8>> = None;
+	let mut cadence: Option<Vec<u8>> = None;
+	let mut metric: Option<Vec<u8>> = None;
+	let mut threshold: Option<Vec<u8>> = None;
 	let mut multipart = Multipart::new(request.body_mut(), boundary);
 	while let Some(field) = multipart.next_field().await? {
 		let name = match field.name() {
@@ -101,7 +103,9 @@ pub async fn post(request: &mut http::Request<hyper::Body>) -> Result<http::Resp
 		};
 		let field_data = field.bytes().await?.to_vec();
 		match name.as_str() {
-			"file" => file = Some(field_data),
+			"cadence" => cadence = Some(field_data),
+			"metric" => metric = Some(field_data),
+			"threshold" => threshold = Some(field_data),
 			_ => {
 				let page = Page {
 					model_layout_info,
@@ -120,47 +124,24 @@ pub async fn post(request: &mut http::Request<hyper::Body>) -> Result<http::Resp
 			}
 		}
 	}
-	let bytes = match file {
-		Some(file) => file,
-		None => {
-			let page = Page {
-				model_layout_info,
-				error: Some("A file is required.".to_owned()),
-			};
-			let html = html(page);
-			let response = http::Response::builder()
-				.status(http::StatusCode::BAD_REQUEST)
-				.body(hyper::Body::from(html))
-				.unwrap();
-			return Ok(response);
-		}
+	// TODO - unwrap options, deal with error, convert Vec<u8> data to proper types
+	let cadence = todo!();
+	let metric = todo!();
+	let variance = todo!();
+	let alert = AlertHeuristics {
+		cadence,
+		thresholds: vec![AlertThreshold { metric, variance }],
 	};
-	let model = match tangram_model::from_bytes(&bytes) {
-		Ok(model) => model,
-		Err(_) => {
-			let page = Page {
-				model_layout_info,
-				error: Some("Invalid tangram model file.".to_owned()),
-			};
-			let html = html(page);
-			let response = http::Response::builder()
-				.status(http::StatusCode::BAD_REQUEST)
-				.body(hyper::Body::from(html))
-				.unwrap();
-			return Ok(response);
-		}
-	};
-	let result = add_model_version(
-		&mut db,
-		&context.storage,
-		repo_id,
-		model.id().parse().unwrap(),
-		&bytes,
+	let result = sqlx::query(
+		"
+			insert into 
+		",
 	)
+	.execute(&mut db)
 	.await;
 	if result.is_err() {
 		let page = Page {
-			model_layout_info, // NEED MODEL_LAYOUT HERE
+			model_layout_info,
 			error: Some("There was an error creating your alert.".to_owned()),
 		};
 		let html = html(page);
@@ -175,11 +156,7 @@ pub async fn post(request: &mut http::Request<hyper::Body>) -> Result<http::Resp
 		.status(http::StatusCode::SEE_OTHER)
 		.header(
 			http::header::LOCATION,
-			format!(
-				"/repos/{}/models/{}/production_alerts/",
-				repo_id,
-				model.id()
-			),
+			format!("/repos/{}/models/{}/production_alerts/", repo_id, model_id),
 		)
 		.body(hyper::Body::empty())
 		.unwrap();
